Support serializing Unicode points properly
While a BuildTree was Unicode-aware before this commit, when serializing
it into a byte slice we assumed one byte per character (rune) which is not
enough to encode all Unicode points.

Instead of using a static entry size which assumes one byte per character
to serialize the character in the node, we use an UTF-8 encoded byte
string of variable size when serializing an entry. The size of the
encoded code point we have to write into the serialization format in
order to be able to read it again. We use three of the yet unused bits
in the flag byte to encode the length of the character byte string in
the node.